Updated package diagram and moved LinkButtons
